fix the format issue, mv spl06 to goertek/spl06
