+++ src/drivers/barometer/goertek/Kconfig
@@ -1,0 +1,3 @@
+menu "Goertek"
+rsource "*/Kconfig"
+endmenu #Goertek
