+++ src/drivers/barometer/goertek/spl06/spl06_main.cpp
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (c) 2016-2019, 2021 PX4 Development Team. All rights reserved.
+ *   Copyright (c) 2022 PX4 Development Team. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -75,6 +75,7 @@ I2CSPIDriverBase *SPL06::instantiate(const I2CSPIDriverConfig &config, int runti
 	}
 
 #endif // CONFIG_SPI
+
 	if (interface == nullptr) {
 		PX4_ERR("failed creating interface for bus %i", config.bus);
 		return nullptr;
@@ -92,6 +93,7 @@ I2CSPIDriverBase *SPL06::instantiate(const I2CSPIDriverConfig &config, int runti
 		delete interface;
 		return nullptr;
 	}
+
 	if (OK != dev->init()) {
 		delete dev;
 		return nullptr;
